Adds utils for better export/import
